update build and transform jest.configs to type commonjs
