+++ package.json
@@ -39,6 +39,5 @@
   },
   "engines": {
     "node": "16.x"
-  },
-  "type":"module"
+  }
 }
